requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
Django==4.1.7
django-cors-headers==3.14.0
django-rest-knox==4.2.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
Flask==2.2.3
Flask-Cors==3.0.10
flask-restplus==0.13.0
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
openapi-codec==1.3.2
packaging==23.0
pycparser==2.21
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
pyrsistent==0.19.3
pytz==2022.7.1
requests==2.28.2
ruamel.yaml==0.17.21
simplejson==3.18.3
six==1.16.0
sqlparse==0.4.3
swagger-ui==0.1.2
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
Werkzeug==2.2.3
